Convert score to float

+++ codeine_django/courses/models.py
@@ -214,7 +214,7 @@ class MRQ(models.Model):
 class QuizResult(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False, unique=True)
     date_created = models.DateTimeField(auto_now_add=True)
-    score = models.PositiveIntegerField(default=0)
+    score = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     passed = models.BooleanField(default=False)
     submitted = models.BooleanField(default=False)
